Check Point
- Moved the AdvancedService from the test_advanced module to a new test utility services module (stackinabox.tests.utils.services)
- Moved HTTPretty tests to their own test module
- Updated to be more Py2/Py3 compliant via using the six module

Note: This versio has an issue with Py2 and PyPy; for some reason it thinks the HelloService was already registered in the Basic HTTPretty test
